backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
sqlalchemy==2.0.30
alembic==1.13.1
aiosqlite==0.20.0
pydantic==2.7.1
pydantic-settings==2.3.0
celery==5.4.0
redis==5.0.6
litellm==1.40.0
langchain==0.2.5
langchain-community==0.2.5
langgraph==0.1.9
sentence-transformers==3.0.1
chromadb==0.5.3
pymupdf==1.24.5
openpyxl==3.1.2
pandas==2.2.2
xlrd==2.0.1
weasyprint==62.3
python-dotenv==1.0.1
httpx==0.27.0
google-generativeai==0.8.3
python-docx==1.1.2
Pillow==10.4.0
